+++ backend/package.json
@@ -11,6 +11,7 @@
     "express": "^4.17.1",
     "firebase-admin": "^9.1.0",
     "morgan": "^1.10.0",
+    "multiparty": "^4.2.2",
     "uuid": "^8.3.0"
   },
   "devDependencies": {
