+++ backend/src/helpers/firebase-admin/index.js
@@ -1,16 +1,15 @@
 import admin from "firebase-admin";
-import "firebase/firebase-storage";
 import env from "../env";
 const serviceAccount = require(env.firebase.serviceAccountPath);
 
 admin.initializeApp({
   credential: admin.credential.cert(serviceAccount),
   databaseURL: env.firebase.dbUrl,
-  storageURL: env.firebase.sUrl
+  storageBucket: env.firebase.sUrl
 });
 
 export const db = admin.firestore();
 
-export const store = firebase.storage;
+export const store = admin.storage().bucket();
 
 export default admin;
