+++ backend/src/index.js
@@ -8,7 +8,7 @@ import authRouter from "./routes/auth-router";
 
 import userRouter from "./routes/user-router";
 
-import uploadRouter from "./routes/upload";
+import uploadRouter from "./routes/upload-router";
 
 import cors from "cors";
 import env from "./helpers/env";
@@ -22,9 +22,13 @@ const config = {
 
 // Middleware
 app.use(morgan("tiny"));
-app.use(cors());
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(bodyParser.json());
+app.use(cors({ limit: '50mb', extended: true }));
+//app.use(bodyParser.urlencoded({ extended: true }));
+app.use(bodyParser.raw({ limit: '10mb', type: 'multipart/form-data' }));
+app.use(bodyParser.json({ limit: '10mb', extended: true }));
+//app.use(bodyParser.urlencoded({ limit: '50mb', extended: true }));
+//app.use(bodyParser.json());
+
 
 // Routes
 app.use("/example", exampleRouter);
