+++ frontend/package.json
@@ -12,6 +12,7 @@
     "axios": "^0.19.2",
     "clsx": "^1.1.0",
     "firebase": "^7.14.5",
+    "form-data": "^3.0.0",
     "material-table": "^1.57.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
