+++ frontend/src/App.js
@@ -65,6 +65,10 @@ function AppRouter(props) {
           <ClassList/>
         </Route>
 
+      <Route path="/upload" exact>
+        <UploadPage />
+      </Route>
+
     </Switch>
   );
 }
@@ -74,7 +78,7 @@ function App() {
     <div className="App">
       <AppProvider>
         <BrowserRouter>
-        <Navigation />
+          <Navigation />
           <div>
             <AppRouter />
             {/* TODO: Fix this so it actually goes to the bottom of the page */}
