+++ frontend/src/helpers/api/upload.js
@@ -2,6 +2,10 @@ import axios from "axios";
 import { API_HOST } from ".";
 
 export const uploadImage = async (fileData) => {
-  const res = await axios.post(`http://${API_HOST}/image/upload`, fileData);
+  const res = await axios.post(`http://${API_HOST}/image/upload`, fileData, {
+    headers: {
+      'Content-Type': 'multipart/form-data'
+    },
+  });
   return res;
 };
