+++ frontend/src/pages/upload/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import UploadImageForm from "../../components/upload"; 
+import UploadImageForm from "../../components/upload_new"; 
 
 export default function testPage() {
   return (
